requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-multipart>=0.0.5
python-dotenv>=0.19.0
requests>=2.31.0
urllib3>=2.0.0
dnspython>=2.1.0
python-whois>=0.7.3
rich>=10.12.0
jinja2>=3.0.1
aiofiles>=0.7.0
python-Wappalyzer>=0.3.0
termcolor>=1.1.0  # Required for EmailHarvester
validators>=0.20.0  # Required for EmailHarvester
aiohttp>=3.9.0

# AI and LLM
langchain>=0.1.0
openai>=1.0.0

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1
pytest-mock>=3.6.1
pytest-asyncio>=0.15.1

# Code quality
flake8>=3.9.2
black>=21.9b0
isort>=5.9.3
mypy>=0.910

# Security
cryptography>=35.0.0
pyOpenSSL>=20.0.1

# Cloud Security
scoutsuite>=5.12.0
boto3>=1.20.0  # For AWS
azure-mgmt-resource==15.0.0  # For Azure (pinned to version compatible with ScoutSuite)
google-cloud-resource-manager>=1.3.3  # For GCP

# Logging and monitoring
prometheus-client>=0.11.0

# PDF Generation
fpdf2>=2.5.4

# These are built-in Python modules, no need to install
# asyncio
# ssl
# concurrent.futures
# sqlite3
# pathlib
# re
# smtplib
# email
# logging
# argparse